extract weight mod.
